score: Remove _Thread_queue_Do_nothing_extract()
This function was unused.  It was a relict of the thread queue rework done
during the SMP support development.  In an early stage, the extract operation
was called with a NULL thread queue.  However, this is no longer the case.  The
extract operation is only called if we have a non-NULL thread queue.